Consider existing auths and captures
When orders that are confirmed are cancelled and brought
back to life as a draft and then processed, existing captures
and authorizations should be respected.

This does not handle an edge case where authorizations are
expired on the payment gateway (after 7 days on stripe) for
example. That is a problem to be solved differently anyway.